[CVW-038] Refactor, 기존 코드 삭제 및 변수명 수정
