--- Projects/Data/Repository/Sources/ToEntityExtension/BinanceOrderbookTableDTO+toEntity.swift
@@ -1,23 +1,0 @@
-//
-//  OrderbookTableDTO+toEntity.swift
-//  Data
-//
-//  Created by choijunios on 4/12/25.
-//
-
-import DataSource
-import DomainInterface
-import CoreUtil
-
-extension BinanceOrderbookTableDTO {
-    func toEntity() -> OrderbookUpdateVO {
-        let converter = { (info: [String]) -> Orderbook in
-            Orderbook(price: CVNumber(Double(info[0])!), quantity: CVNumber(Double(info[1])!))
-        }
-        return OrderbookUpdateVO(
-            bids: bids.map(converter),
-            asks: asks.map(converter),
-            lastUpdateId: self.lastUpdateId
-        )
-    }
-}
