+++ Projects/Domain/Interface/Entity/Orderbook/Orderbook.swift
@@ -1,0 +1,17 @@
+//
+//  Orderbook.swift
+//  Domain
+//
+//  Created by choijunios on 4/21/25.
+//
+
+import CoreUtil
+
+public struct Orderbook: Sendable {
+    public let price: CVNumber
+    public let quantity: CVNumber
+    public init(price: CVNumber, quantity: CVNumber) {
+        self.price = price
+        self.quantity = quantity
+    }
+}
